% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioclimatic}
\alias{bioclimatic}
\title{Bioclimatic variables}
\format{
An object of class \code{data.frame} with 1862 rows and 14 columns.
}
\usage{
bioclimatic
}
\description{
A data set of bioclimatic variables for \eqn{n=1,862} locations homogeneously distributed over the global landmass from CHELSA("Climatologies at high resolution for the earth’s land surface areas").
}
\examples{
head(bioclimatic)
}
\references{
D.N. Karger, O. Conrad, J. Böhner, T. Kawohl, H. Kreft, R.W. Soria-Auza, N.E. Zimmermann, H.P. Linder, M. Kessler, Climatologies at high resolution for the Earth's land surface areas, Sci. Data 4(1), 2017.
}
\keyword{datasets}
