% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.concor.M}
\alias{clust.concor.M}
\title{Clustering a set of variables using distance function based on multivariate concordance measures}
\usage{
clust.concor.M(X, method = c("footrule"))
}
\arguments{
\item{X}{a data frame for vector X}

\item{method}{kendall / footrule}
}
\value{
a list for hierarchical clustering result
}
\description{
Clustering a set of variables using distance function based on multivariate concordance measures
}
\keyword{internal}
