% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.Tq}
\alias{clust.Tq}
\title{Cluster a set of variables using distance function based on predictive measure}
\usage{
clust.Tq(X, estim.method = c("copula"), mutual = FALSE)
}
\arguments{
\item{X}{a data frame for a set of variables X}

\item{estim.method}{An optional character string specifying a method for estimating the directed dependence coefficient.}

\item{mutual}{type B function or not}
}
\value{
a list for hierarchical clustering result
}
\description{
Cluster a set of variables using distance function based on predictive measure
}
\keyword{internal}
