% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{Codec.Tq.Perm}
\alias{Codec.Tq.Perm}
\title{Estimate for T^q_bar(Y|X) based on function Codec & a sample of all / all increasing / all decreasing permutations}
\usage{
Codec.Tq.Perm(X, Y, method = c("sample"))
}
\arguments{
\item{X}{a data frame for input vector X}

\item{Y}{a data frame for output vector Y}

\item{method}{permuatation methods: sample / increasing / decreasing / full}
}
\value{
the value of T^q_bar(Y|X)
}
\description{
Estimate for T^q_bar(Y|X) based on function Codec & a sample of all / all increasing / all decreasing permutations
}
\keyword{internal}
