% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{Codec.Tq}
\alias{Codec.Tq}
\title{Estimate for T^q(Y|X) based on function Codec}
\usage{
Codec.Tq(X, Y)
}
\arguments{
\item{X}{a data frame for input vector X}

\item{Y}{a data frame for output vector Y}
}
\value{
the value of T^q(Y|X)
}
\description{
Estimate for T^q(Y|X) based on function Codec
}
\keyword{internal}
