% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{CodecCorr}
\alias{CodecCorr}
\title{Estimate for \eqn{\xi}(Y,X) using codec function}
\usage{
CodecCorr(X, y)
}
\arguments{
\item{X}{a data frame for input vector X}

\item{y}{a data frame for output vector Y}
}
\value{
a value
}
\description{
Estimate for \eqn{\xi}(Y,X) using codec function
}
\keyword{internal}
