% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{concor.M}
\alias{concor.M}
\title{Estimation for multivariate concordance measures}
\usage{
concor.M(X, method = c("footrule"))
}
\arguments{
\item{X}{a data frame for vector X}

\item{method}{kendall / footrule}
}
\value{
a value of the estimator for the multivariate concordance measures
}
\description{
Estimation for multivariate concordance measures
}
\keyword{internal}
