% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{Copula.Tq.Perm}
\alias{Copula.Tq.Perm}
\title{Estimate for T_bar(Y,X) based on dimension reduction principle}
\usage{
Copula.Tq.Perm(X, Y, method = c("sample"))
}
\arguments{
\item{X}{a data frame for input vector X}

\item{Y}{a data frame for output vector Y}

\item{method}{permuatation methods: sample / increasing / decreasing / full}
}
\value{
a value
}
\description{
Estimate for T_bar(Y,X) based on dimension reduction principle
}
\keyword{internal}
