% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{Copula.Tq}
\alias{Copula.Tq}
\title{Estimate for T(Y,X) based on dimension reduction principle}
\usage{
Copula.Tq(X, Y)
}
\arguments{
\item{X}{a data frame for input vector X}

\item{Y}{a data frame for output vector Y}
}
\value{
a value
}
\description{
Estimate for T(Y,X) based on dimension reduction principle
}
\keyword{internal}
