% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{CopulaCorr}
\alias{CopulaCorr}
\title{Estimate for \eqn{\xi}(Y,X) based on dimension reduction principle}
\usage{
CopulaCorr(X, y)
}
\arguments{
\item{X}{a data frame for input vector X}

\item{y}{a data frame for output vector Y}
}
\value{
a value
}
\description{
Estimate for \eqn{\xi}(Y,X) based on dimension reduction principle
}
\keyword{internal}
