% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{dendrogram}
\alias{dendrogram}
\title{Read a dendrogram from a list for hierarchical clustering result}
\usage{
dendrogram(clust, step = TRUE)
}
\arguments{
\item{clust}{a list for hierarchical clustering result}

\item{step}{whether using clustering step as y axis or not}
}
\value{
an object of class "dendrogram"
}
\description{
Read a dendrogram from a list for hierarchical clustering result
}
\keyword{internal}
