% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.R
\name{dist.concor.M}
\alias{dist.concor.M}
\title{distance function based on multivariate concordance measures}
\usage{
dist.concor.M(X, Y, method = c("footrule"))
}
\arguments{
\item{X}{a data frame for vector X}

\item{Y}{a data frame for vector Y}

\item{method}{kendall / footrule}
}
\value{
a value for distance between two vectors
}
\description{
distance function based on multivariate concordance measures
}
\keyword{internal}
