% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.R
\name{dist.mat.concor}
\alias{dist.mat.concor}
\title{Distance Matrix Computation using distance function based on multivariate concordance measures}
\usage{
dist.mat.concor(X, method = c("footrule"))
}
\arguments{
\item{X}{a data frame for vector X}

\item{method}{kendall / footrule}
}
\value{
an object of class "dist"
}
\description{
Distance Matrix Computation using distance function based on multivariate concordance measures
}
\keyword{internal}
