% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.R
\name{dist.mat.T}
\alias{dist.mat.T}
\title{Distance Matrix Computation using distance function based on T^q}
\usage{
dist.mat.T(X, mutual = FALSE)
}
\arguments{
\item{X}{a data frame for vector X}

\item{mutual}{use type B function (mutual perfect dependence) or not}
}
\value{
an object of class "dist"
}
\description{
Distance Matrix Computation using distance function based on T^q
}
\keyword{internal}
