% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.R
\name{dist.Tq}
\alias{dist.Tq}
\title{distance function based on T}
\usage{
dist.Tq(X, Y, estim.method = c("copula"), mutual = FALSE)
}
\arguments{
\item{X}{a data frame for vector X}

\item{Y}{a data frame for vector Y}

\item{estim.method}{An optional character string specifying a method for estimating the directed dependence coefficient.}

\item{mutual}{use mutual perfect dependence or not}
}
\value{
a value for distance between two vectors
}
\description{
distance function based on T
}
\keyword{internal}
