% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{MPhi}
\alias{MPhi}
\title{Markov product estimate from single (q=1) endogenuous and (p>=1) exogenous variables based on dimension reduction}
\usage{
MPhi(X, y)
}
\arguments{
\item{X}{a data frame for input vector X}

\item{y}{a data frame for output vector Y}
}
\value{
a value
}
\description{
Markov product estimate from single (q=1) endogenuous and (p>=1) exogenous variables based on dimension reduction
}
\keyword{internal}
