% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Adiam.Msplit}
\alias{plot_Adiam.Msplit}
\title{Plot the trade-off of Adiam and Msplit}
\usage{
plot_Adiam.Msplit(tradeoff, main = NULL, sub = NULL)
}
\arguments{
\item{tradeoff}{a data frame}

\item{main}{main title}

\item{sub}{sub title}
}
\value{
ggplot
}
\description{
Plot the trade-off of Adiam and Msplit
}
\keyword{internal}
