% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{plotting a dendrogram with colored branches}
\usage{
plot_dendrogram(
  dend,
  num.cluster = num.cluster,
  linkage = FALSE,
  ylab = ylab,
  cex.lab = 0.6,
  cex.axis = 0.6
)
}
\arguments{
\item{dend}{a dendrogram}

\item{num.cluster}{the number of colored branches}

\item{linkage}{logical; if 'True', the linkage method is used}

\item{ylab}{a string}

\item{cex.lab}{a value}

\item{cex.axis}{a value}
}
\value{
plot
}
\description{
plotting a dendrogram with colored branches
}
\keyword{internal}
