% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Silhouette.coefficient}
\alias{plot_Silhouette.coefficient}
\title{Plot the Silhouette coefficient}
\usage{
plot_Silhouette.coefficient(Silhouette_Index, main = NULL, sub = NULL)
}
\arguments{
\item{Silhouette_Index}{a data frame of Silhouette coefficient}

\item{main}{main title}

\item{sub}{sub title}
}
\value{
ggplot
}
\description{
Plot the Silhouette coefficient
}
\keyword{internal}
