% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.partition.R
\name{Silhouette}
\alias{Silhouette}
\title{Silhouette value for the i. variable given variable partition}
\usage{
Silhouette(i, df, partition, dist.func = "PD", estim.method = c("copula"))
}
\arguments{
\item{i}{the index of the variable}

\item{df}{a data frame for all variables}

\item{partition}{a partition}

\item{dist.func}{PD / MPD / kendall / footrule}

\item{estim.method}{An optional character string specifying a method for estimating the directed dependence coefficient.}
}
\value{
a value for Silhouette
}
\description{
Silhouette value for the i. variable given variable partition
}
\keyword{internal}
