% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.partition.R
\name{split}
\alias{split}
\title{Split of two classes of variables based on different distance function}
\usage{
split(X, Y, dist.func = "PD", estim.method = c("copula"))
}
\arguments{
\item{X}{a data frame for a set of variables X}

\item{Y}{a data frame for a set of variables Y}

\item{dist.func}{PD / MPD / kendall / footrule}

\item{estim.method}{An optional character string specifying a method for estimating the directed dependence coefficient.}
}
\value{
a value
}
\description{
Split of two classes of variables based on different distance function
}
\keyword{internal}
