% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.Tq}
\alias{clust.Tq}
\title{Cluster a set of variables using distance function based on predictive measure}
\usage{
clust.Tq(X, perm = TRUE, perm.method = c("decreasing"), mutual = FALSE)
}
\arguments{
\item{X}{a data frame for a set of variables X}

\item{perm}{T^q or T^q_bar}

\item{perm.method}{permutation methods: sample / increasing / decreasing}

\item{mutual}{type B function or not}
}
\value{
a list for hierarchical clustering result
}
\description{
Cluster a set of variables using distance function based on predictive measure
}
\keyword{internal}
