% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.R
\name{dist.Tq}
\alias{dist.Tq}
\title{distance function based on T^q}
\usage{
dist.Tq(X, Y, perm = TRUE, perm.method = c("decreasing"), mutual = FALSE)
}
\arguments{
\item{X}{a data frame for vector X}

\item{Y}{a data frame for vector Y}

\item{perm}{permuted version or not}

\item{perm.method}{permutation methods: sample / increasing / decreasing / full}

\item{mutual}{use mutual perfect dependence or not}
}
\value{
a value for distance between two vectors
}
\description{
distance function based on T^q
}
\keyword{internal}
